+++ pkg/runtimeoptions/v1/api.proto
@@ -1,7 +1,7 @@
 // To regenerate api.pb.go run `make protos`
 syntax = "proto3";
 
-package cri.runtimeoptions.v1;
+package runtimeoptions.v1;
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 
@@ -14,7 +14,7 @@ option (gogoproto.unmarshaler_all) = true;
 option (gogoproto.goproto_unrecognized_all) = false;
 
 
-option go_package = "github.com/containerd/containerd/pkg/cri/runtimeoptions/v1;cri_runtimeoptions_v1";
+option go_package = "github.com/containerd/containerd/pkg/runtimeoptions/v1;runtimeoptions_v1";
 
 message Options {
 	// TypeUrl specifies the type of the content inside the config file.
